requirements.txt


# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database & ORM
prisma==0.15.0
psycopg2-binary==2.9.10

# Settings & Configuration
pydantic-settings==2.0.3
python-dotenv==1.0.0

# API & Security
slowapi==0.1.9
python-multipart==0.0.6

# Data Processing (with compatible versions)
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
openpyxl==3.1.2
xlsxwriter==3.1.9

# HTTP Client
httpx==0.24.1

# Production Server
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
